regression: error when Dir.home is not readable
When ``Dir.home`` is not readable ``Netrc.read`` throws an exception [at
least on linux]. There was an failed attempt to fix this issue in the
past [``File.stat.readable?`` could also throw] but at some point an
early return has been added that returned the content of ``Dir.home``
when that was defined, completely skipping the path validation.

The new code always checks that the path is readable [according to
``File.readable?``] before returning it [linux only].

Lastly, ``"./"`` has been replaced by ``Dir.pwd``, which should produce
the same result.