fixes #46 (unit tests): missing $HOME relevant only w/o passwd->pw_dir
When Ruby's `Dir` class obtains the user's home directory from the password
database via `getpwname(3)`, the `${HOME}` environment variable is not
consulted. In such a case, `Dir.home` will return the value from the `passwd`
struct's `pw_dir` member, and the last resort fallback use of `Dir.pwd` (by
`Netrc.home_path`) is not triggered.

The `test_missing_environment()` unit test was failing on my machine (Debian
GNU/Linux, amd64, libc6 2.28, Ruby 2.5) because it expected that by removing
the `${HOME}` environment variable the `Netrc` class would have constructed
its `default_path` from the `Dir.pwd` fallback. That was not the behavior,
though, when Ruby's `Dir.home` was able to return a valid, readable directory
from the password database (and the build process was not in that directory
when running the unit tests).

This change tweaks the unit test to conditionally expect the use of `Dir.home`
rather than `Dir.pwd`.