fixes #50: augment Dir.home to read password database by uid
The `Dir.home` module present in versions of Ruby 2.x through the latest
(2.7.1, released 2020-03-31) is unable to reliably locate the user's home
directory when all three of the following are true at the same time:

   1. Ruby is running on a Unix-like OS
   2. The `$HOME` environment variable is not set
   3. The process is not a descendant of login(1) (or a work-alike)

On a Unix-like OS, when the $HOME environment variable is not set, one might
expect that Ruby would attempt to obtain the user's home directory from the
password database. And in fact, it does try to do so. But the mechanism it
uses only works for (grand)children of `login(1)`. In particular, it uses
`getlogin(3)` to obtain the username, with the intent to then obtain the
user's password record (and its `pw_dir` member) by looking it up by name
(`getpwnam(3)`). That `getlogin()` call fails, of course, because there is no
logged-in user for the process.

This change preserves the base level reliance on `Dir.home`, but augments it
by catching the error it throws in the above scenario and then using the Ruby
stdlib wrappers for `getuid(1)` and `getpwuid(3)` to obtain the user's record
from the password database. This works reliably and the same regardless of
whether or not the process is a (grand)child of login.

This change should be forward compatible with future versions of Ruby should
the `Dir.home` module adapt the approach we use here. In such a case, the
module would cease to throw the exception that it does currently, and this
augmenting code will not be executed.

This change also simplifies the `test_missing_environment` unit test.

This issue surfaced when integration tests were first setup for the `Netrc`
project on GitHub Actions. The process that runs the unit tests there is not a
(grand)child of login(1), so fails on the unit tests that exercise the "what
to do when $HOME is unset" logic.